feat: 完成用户认证系统和项目文档 - 实现完整的用户认证系统和Flutter应用基础架构 - 修复登录API响应格式和JSON序列化问题 - 添加完整的项目README.md和MIT许可证 - 建立微服务架构和数据库配置

+++ mobile/lib/features/analytics/presentation/widgets/performance_metrics_card.dart
@@ -46,7 +46,7 @@ class PerformanceMetricsCard extends StatelessWidget {
                 context,
                 '平均响应时间',
                 '${performanceData!.averageResponseTime.toStringAsFixed(0)}ms',
-                _getResponseTimeColor(performanceData!.averageResponseTime),
+                _getResponseTimeColor(performanceData!.averageResponseTime.toDouble()),
               ),
               const SizedBox(height: 8),
